requirements.txt
fastapi
uvicorn
opencv-python
pillow
pytesseract
numpy
python-multipart